option markup rake task
rake take to help build markup

`rake dev:option_markup` will run a wizard to that will build the option markup for the edit and index pages.

currently this will just return the markup in the terminal, so it will have to be copied and pasted to a widget.
eventually it would be good to append this to the appropriate files.