Make fallback lead text show by default.
Originally we had the text in noscript, and for browsers without CORS
but with JS we would redirect them. That didn't work so hot in practice
because the script is on every page. We removed the CORS check, but that
means some browsers, particularly IE, does part of the job (whiping the
div) but doesn't actually replace it because it blows up.

So show this text. Hopefully it doesn't flash on capable browsers.