Bring back CORS check for IUI.
This got dropped because it was a temporary hack that accidentally got
lost when the IUI build script was subsequently re-ran. I have updated
the build script, but I need to add that hack back to the built widget.
Also it's not a hack anymore, it's the permanent solution and I like it.
No IE9, no problem.